refactor: Remove redundant methods from CRUDEvent
The methods were redundant because the parent class already implements them.
Removing them also fixed an error where the method remove called the same method in the parent class with a wrong parameter name.

To pass the owner_uuid, a new Pydantic model was implemented, separate from the model used in the router to validate requests.